fix: move interface to model and add useFormContext
