+++ src/features/AddressForm/model/types.ts
@@ -1,0 +1,6 @@
+export interface AddressFormValues {
+  country: string;
+  city: string;
+  streetName: string;
+  postalCode: string;
+}
